Weird API stuff MK IV.

+++ src/main/java/Controllers/TaskManagement.java
@@ -1,59 +1,105 @@
 package Controllers;
 
 import Server.Main;
+import org.glassfish.jersey.media.multipart.FormDataParam;
+import org.json.simple.JSONArray;
+import org.json.simple.JSONObject;
 
+import javax.ws.rs.*;
+import javax.ws.rs.core.MediaType;
 import java.sql.ResultSet;
 import java.sql.PreparedStatement;
 
+@Path("Tasks/")
 public class TaskManagement {
-    public static void addTask(Integer taskID, String task, String answer) {
+    @POST
+    @Path("addTask")
+    @Consumes(MediaType.MULTIPART_FORM_DATA)
+    @Produces(MediaType.APPLICATION_JSON)
+    public String addTask(@FormDataParam("taskID")Integer taskID,
+                          @FormDataParam("task")String task,
+                          @FormDataParam("answer")String answer,
+                          @FormDataParam("toolID")Integer toolID) {
         try {
-            PreparedStatement ps = Main.db.prepareStatement("INSERT INTO Tasks(TaskID, Task, CorrectAnswer) VALUES (?, ?, ?)");
+            if (taskID == null || task == null || answer == null||toolID  == null) {
+                throw new Exception("One or more form data parameters are missing in the HTTP request.");
+            }
+            PreparedStatement ps = Main.db.prepareStatement("INSERT INTO Tasks(TaskID, Task, CorrectAnswer, ToolID) VALUES (?, ?, ?, ?)");
             ps.setInt(1, taskID);
             ps.setString(2, task);
             ps.setString(3, answer);
-            ps.executeUpdate();
+            ps.setInt(4, toolID);
+            ps.execute();
+            return "{\"status\": \"OK\"}";
 
         } catch (Exception exception) {
             System.out.println("Database disconnection error: " + exception.getMessage());
+            return "{\"error\": \"Unable to create new item, please see server console for more info.\"}";
         }
     }
 
-    public static void readTask() {
+    @GET
+    @Path("list")
+    @Produces(MediaType.APPLICATION_JSON)
+    public String readTask() {
+        System.out.println("thing/list");
+        JSONArray list = new JSONArray();
         try {
 
             PreparedStatement ps = Main.db.prepareStatement("SELECT Task FROM Tasks");
             ResultSet results = ps.executeQuery();
             while (results.next()) {
-                String task = results.getString(1);
-                System.out.println(task);
+                JSONObject item = new JSONObject();
+                item.put("name", results.getString(1));
+                list.add(item);
             }
+            return list.toString();
 
 
         } catch (Exception exception) {
             System.out.println("Database disconnection error: " + exception.getMessage());
+            return "{\"error\": \"Unable to list items, please see server console for more info.\"}";
         }
     }
 
-    public static void updateTask(String task, Integer taskID) {
+    @POST
+    @Path("updateTask")
+    @Consumes(MediaType.MULTIPART_FORM_DATA)
+    @Produces(MediaType.APPLICATION_JSON)
+    public String updateTask(@FormDataParam("task")String task,
+                             @FormDataParam("taskID")Integer taskID) {
         try {
+            if (task == null || taskID == null) {
+                throw new Exception("One or more form data parameters are missing in the HTTP request.");
+            }
             PreparedStatement ps = Main.db.prepareStatement("UPDATE Tasks SET Task = ? WHERE TaskID = ?");
             ps.setString(1, task);
             ps.setInt(2, taskID);
             ps.executeUpdate();
+            return "{\"status\": \"OK\"}";
 
         } catch (Exception exception) {
             System.out.println("Database disconnection error: " + exception.getMessage());
+            return "{\"error\": \"Unable to update item, please see server console for more info.\"}";
         }
     }
-    public static void deleteTask(Integer taskID) {
+    @POST
+    @Path("deleteTask")
+    @Consumes(MediaType.MULTIPART_FORM_DATA)
+    @Produces(MediaType.APPLICATION_JSON)
+    public String deleteTask(@FormDataParam("taskID")Integer taskID) {
         try {
+            if (taskID == null) {
+                throw new Exception("One or more form data parameters are missing in the HTTP request.");
+            }
             PreparedStatement ps = Main.db.prepareStatement("DELETE FROM Tasks WHERE TaskID = ?");
             ps.setInt(1, taskID);
             ps.executeUpdate();
+            return "{\"status\": \"OK\"}";
 
         } catch (Exception exception) {
             System.out.println("Database disconnection error: " + exception.getMessage());
+            return "{\"error\": \"Unable to delete item, please see server console for more info.\"}";
         }
     }
 }
